+++ bui-demo/frontend_rust_wasm/build.sh
@@ -1,0 +1,8 @@
+#!/bin/bash -x
+set -o errexit
+
+cargo web build --target-webasm --release
+
+mkdir -p dist
+cp -a target/wasm32-unknown-unknown/release/bui-demo-frontend-rust-wasm.* dist/
+cp -a static/* dist/